requirements.txt
flet==0.25.1
pdfplumber
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0
regex>=2023.10.3